fix(review): eliminate VM imports from UI component tree and fix lints
