+++ lib/app.dart
@@ -2,22 +2,24 @@ import 'package:jaspr/dom.dart';
 import 'package:jaspr/jaspr.dart';
 import 'package:jaspr_router/jaspr_router.dart';
 
-import 'content.dart' as content;
+import 'models/data_model.dart';
 import 'pages/about.dart';
 import 'pages/home.dart';
 import 'pages/post_page.dart';
 import 'pages/projects_page.dart';
 import 'pages/static_outputs.dart';
-import 'projects_content.dart' as projects_content;
 
 class App extends StatelessComponent {
-  const App({super.key});
+  final List<Post> posts;
+  final List<Project> projects;
+
+  const App({required this.posts, required this.projects, super.key});
 
   @override
   Component build(BuildContext context) {
     // Dynamically compile post routes from our parsed content database
-    final postRoutes = content.posts
-        .where((post) => post.flavor == content.EntryFlavor.writing)
+    final postRoutes = posts
+        .where((post) => post.flavor == EntryFlavor.writing)
         .map(
           (post) => Route(
             path: post.permalink,
@@ -33,7 +35,7 @@ class App extends StatelessComponent {
           Route(
             path: '/',
             title: 'kevmoo @ Work',
-            builder: (context, state) => const Home(),
+            builder: (context, state) => Home(posts: posts),
           ),
           Route(
             path: '/about',
@@ -43,8 +45,7 @@ class App extends StatelessComponent {
           Route(
             path: '/projects',
             title: 'Projects | kevmoo @ Work',
-            builder: (context, state) =>
-                ProjectsPage(projects: projects_content.projects),
+            builder: (context, state) => ProjectsPage(projects: projects),
           ),
           Route(
             path: '/feed.xml',
