+++ lib/models/data_model.dart
@@ -136,3 +136,16 @@ const socialLinks = <SocialLink>[
     iconClass: 'fab fa-twitter',
   ),
 ];
+
+String normalizeTag(String tag) => switch (tag.toLowerCase()) {
+  'dartlang' || 'dart' => 'Dart',
+  'ruby on rails' || 'rails' => 'Rails',
+  '.net' => '.NET',
+  'c#' => 'CSharp',
+  'wpf' => 'WPF',
+  'xaml' => 'XAML',
+  'git' => 'Git',
+  'javascript' => 'JS',
+  _ when tag.length <= 3 => tag.toUpperCase(),
+  _ => tag[0].toUpperCase() + tag.substring(1),
+};
